+++ server-sdk/src/main/java/com/inductiveautomation/opcua/sdk/server/api/Namespace.java
@@ -16,9 +16,11 @@
 
 package com.inductiveautomation.opcua.sdk.server.api;
 
+import com.inductiveautomation.opcua.stack.core.types.builtin.unsigned.UShort;
+
 public interface Namespace extends AttributeManager, MethodManager, MonitoredItemManager, NodeManager {
 
-    int getNamespaceIndex();
+    UShort getNamespaceIndex();
 
     String getNamespaceUri();
